Deal with rasp pi losing internet connection.
Some of the raspberry pis seem to lose their WiFi connection. Google
implies that this is a somewhat common problem. My experience is that
20% of the devices are hit with this problem and they may develop it
after months of working just fine.

To deal with this, the simplest solution is to reboot. We'll check
NOAA a few times before rebooting. This is for allowing initial device
setup to occur without a reboot loop.